requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.17
djangorestframework-simplejwt>=5.5
django-cors-headers>=4.7
django-filter>=25.1
drf-spectacular>=0.28
Pillow>=12.0
python-decouple>=3.8
PyYAML>=6.0

